--- src/pages/Dashboard/Actions/index.js
@@ -1,6 +1,0 @@
-export const createTimeEntry = (timeEntry) => {
-  return {
-    type: 'CREATE_TIME_ENTRY',
-    timeEntry
-  }
-}
