--- src/pages/Login/Actions/index.js
@@ -1,13 +1,0 @@
-export const doLogin = (credentials) => {
-  return {
-    type: 'DO_LOGIN',
-    credentials
-  }
-}
-
-export const attemptLogin = (credentials) => {
-  return function(dispatch, getState) {
-    const state = getState();
-    dispatch(doLogin(credentials));
-  }
-}
