--- src/pages/Projects/Actions/index.js
@@ -1,6 +1,0 @@
-export const createProject = (project) => {
-  return {
-    type: 'CREATE_PROJECT',
-    project
-  }
-}
